Change account server certificate feature to theme 2
MainTheme.colors.onBackgroundSecondary does not exist and needs a design review
