+++ feature/account/server/certificate/src/main/kotlin/app/k9mail/feature/account/server/certificate/ui/ServerCertificateErrorContent.kt
@@ -9,15 +9,11 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.requiredSize
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import app.k9mail.core.ui.compose.common.baseline.withBaseline
-import app.k9mail.core.ui.compose.common.koin.koinPreview
 import app.k9mail.core.ui.compose.common.resources.annotatedStringResource
 import app.k9mail.core.ui.compose.common.text.bold
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
@@ -26,13 +22,10 @@ import app.k9mail.core.ui.compose.designsystem.atom.text.TextBodyLarge
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextHeadlineMedium
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextTitleMedium
 import app.k9mail.core.ui.compose.designsystem.template.ResponsiveWidthContainer
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.MainTheme
+import app.k9mail.core.ui.compose.theme2.MainTheme
 import app.k9mail.feature.account.server.certificate.R
 import app.k9mail.feature.account.server.certificate.domain.entity.FormattedServerCertificateError
-import app.k9mail.feature.account.server.certificate.domain.entity.ServerCertificateProperties
 import app.k9mail.feature.account.server.certificate.ui.ServerCertificateErrorContract.State
-import okio.ByteString.Companion.decodeHex
 import org.koin.compose.koinInject
 
 @Composable
@@ -112,41 +105,4 @@ private fun CertificateErrorDescription(
             serverNameFormatter.format(certificateError.hostname).bold(),
         ),
     )
-}
-
-@Composable
-@Preview(showBackground = true)
-internal fun ServerCertificateErrorContentPreview() {
-    val state = State(
-        isShowServerCertificate = true,
-        certificateError = FormattedServerCertificateError(
-            hostname = "mail.domain.example",
-            serverCertificateProperties = ServerCertificateProperties(
-                subjectAlternativeNames = listOf("*.domain.example", "domain.example"),
-                notValidBefore = "January 1, 2023, 12:00 AM",
-                notValidAfter = "December 31, 2023, 11:59 PM",
-                subject = "CN=*.domain.example",
-                issuer = "CN=test, O=MZLA",
-                fingerprintSha1 = "33ab5639bfd8e7b95eb1d8d0b87781d4ffea4d5d".decodeHex(),
-                fingerprintSha256 = "1894a19c85ba153acbf743ac4e43fc004c891604b26f8c69e1e83ea2afc7c48f".decodeHex(),
-                fingerprintSha512 = (
-                    "81381f1dacd4824a6c503fd07057763099c12b8309d0abcec4000c9060cbbfa6" +
-                        "7988b2ada669ab4837fcd3d4ea6e2b8db2b9da9197d5112fb369fd006da545de"
-                    ).decodeHex(),
-            ),
-        ),
-    )
-
-    koinPreview {
-        factory<ServerNameFormatter> { DefaultServerNameFormatter() }
-        factory<FingerprintFormatter> { DefaultFingerprintFormatter() }
-    } WithContent {
-        K9Theme {
-            ServerCertificateErrorContent(
-                innerPadding = PaddingValues(all = 0.dp),
-                state = state,
-                scrollState = rememberScrollState(),
-            )
-        }
-    }
 }
